Module cordyceps::list

An intrusive doubly-linked list.

See the List type for details.

Structs

Cursor

A cursor over a List.

CursorMut

A cursor over a List with editing operations.

DrainFilter

An iterator returned by List::drain_filter.

IntoIter

An owning iterator over the elements of a List.

Iter

Iterates over the items in a List by reference.

IterMut

Iterates over the items in a List by mutable reference.

Links

Links to other nodes in a List.

List

An intrusive doubly-linked list.