Module cordyceps::mpsc_queue

A multi-producer, single-consumer (MPSC) queue, implemented using a lock-free intrusive singly-linked list.

See the documentation for the MpscQueue type for details.

Based on Dmitry Vyukov’s intrusive MPSC.

Structs

Consumer

A handle that holds the right to dequeue elements from a MpscQueue.

Links

Links to other nodes in a MpscQueue.

MpscQueue

A multi-producer, single-consumer (MPSC) queue, implemented using a lock-free intrusive singly-linked list.

OwnedConsumer

An owned handle that holds the right to dequeue elements from the queue.

Enums

TryDequeueError

Errors returned by MpscQueue::try_dequeue and Consumer::try_dequeue.